src/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
click==8.1.3
Django==4.1.1
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
whitenoise==6.2.0
